Important detail
Upon reviewing the phenotypic data file, I've noticed an inconsistency in the definitions. When DX_GROUP = 1 (indicating autism), the label is set as 0, whereas in the README file, 0 is labeled as "negative" (non-afflicted). For instance, for patient 50643 (male, 21 years old, diagnosed with autism), the label in abide.npy is 0 (negative).